refactored ShowcaseGrid to create ProjectsGrid for reuse on projects page
